+++ lib/JavaScript.ml
@@ -3,19 +3,23 @@
 let fprintf = Format.fprintf
 
 let rec print_program ppf program =
-  match SWIR.program_constants program with
+  begin match SWIR.program_constants program with
   | [] -> ()
   | constants -> begin
       Format.pp_print_list ~pp_sep:Format.pp_print_cut print_constant ppf constants;
       Format.pp_print_cut ppf ()
-    end;
-  match SWIR.program_functions program with
+    end
+  end;
+  begin match SWIR.program_functions program with
   | [] -> ()
   | functions -> begin
       Format.pp_print_list ~pp_sep:Format.pp_print_cut print_function ppf functions;
       Format.pp_print_cut ppf ()
-    end;
-  fprintf ppf "@[<v 2>export function handle(state, action) {@,%a@]@,}@." print_handle_function program
+    end
+  end;
+  fprintf ppf "@[<v 2>export function handle(state, action) {@,";
+  fprintf ppf "const input = action.input@,";
+  fprintf ppf "%a@]@,}@." print_handle_function program
 
 and print_constant ppf = function
   | SWIR.Const (name, _, value) -> fprintf ppf "const %s = %a@," name print_expression value
@@ -23,27 +27,44 @@ and print_constant ppf = function
 
 and print_handle_function ppf program =
   let functions = SWIR.program_functions program in
+  let filter = function
+    | (SWIR.Function (SWIR.Public, _, _, _)) as f -> Some f
+    | (SWIR.Function (SWIR.PublicPure, _, _, _)) as f -> Some f
+    | _ -> None
+  in
+  let functions = List.filter_map filter functions in
   Format.pp_print_list ~pp_sep:Format.pp_print_cut print_handle_clause ppf functions;
   fprintf ppf "@,return {state}"
 
 and print_handle_clause ppf = function
-  | Function (id, _, _) ->
+  | Function (PublicPure, id, _, _) ->
     let js_name = mangle_name id in
-    fprintf ppf "@[<v 2>if (action.input.function === '%s') {@,%s(state)@,return {state}@]@,}" js_name js_name
-    | _ -> failwith "unreachable"
+    fprintf ppf "@[<v 2>if (input.function === '%s') {@,return {result: %s(state)}@]@,}" js_name js_name
+  | Function (Public, id, _, _) ->
+    let js_name = mangle_name id in
+    fprintf ppf "@[<v 2>if (input.function === '%s') {@,return {state: %s(state)}@]@,}" js_name js_name
+  | _ -> failwith "unreachable"
 
 and print_function ppf = function
-  | SWIR.Function (name, params, body) ->
+  | SWIR.Function (modifier, name, params, body) ->
     let params = ("state", None) :: params in
-    fprintf ppf "@[<v 2>function %s(@[<h>%a@]) {@,%a@]@,}@,"
-      (mangle_name name) print_parameters params print_expressions body
+    fprintf ppf "@[<v 2>function %s(@[<h>%a@]) {@,%a"
+      (mangle_name name)
+      print_parameters params
+      print_expressions body;
+    begin match modifier with
+      | Public -> fprintf ppf "@,return state"
+      | _ -> fprintf ppf "@,return result"
+    end;
+    fprintf ppf "@]@,}@,"
   | _ -> failwith "unreachable"
 
 and print_parameters ppf params =
-  Format.pp_print_list ~pp_sep:Format.pp_print_space print_parameter ppf params
+  let print_comma ppf () = Format.fprintf ppf ",@ " in
+  Format.pp_print_list ~pp_sep:print_comma print_parameter ppf params
 
 and print_parameter ppf = function
-  | (name, _) -> fprintf ppf "%s" name
+  | (name, _) -> fprintf ppf "%s" (mangle_name name)
 
 and print_expressions ppf = function
   | [expr] -> fprintf ppf "%a" print_expression expr
@@ -53,7 +74,7 @@ and print_expression ppf = function
   | SWIR.Literal lit -> print_literal ppf lit
   | SWIR.VarGet var -> fprintf ppf "state.%s" var
   | SWIR.VarSet (var, val') -> fprintf ppf "state.%s = %a" var print_expression val'
-  | SWIR.Ok expr -> fprintf ppf "return %a" print_expression expr
+  | SWIR.Ok expr -> fprintf ppf "const result = %a" print_expression expr
   | SWIR.Add (a, b) ->
     fprintf ppf "(%a + %a)" print_expression a print_expression b
   | SWIR.Sub (a, b) ->